Removes System.Drawing from IdSharp.Tagging
Coupling this library with System.Drawing for one use case is unnecessary, especially when many libraries have gone to great pains to remove System.Drawing form their stacks.

This commit moves the System.Drawing parts of AttachedPicture into a separate project  so that those references do not pollute the main library.

More work is needed to, especially in deciding whether or not to keep the AttachedPictureWithImage class.