Linux Example Project (#25)
* Created a Linux example project.
The Linux project can work only with version 6.0.0 of System.Drawing.Common. The support for this was removed in version 7 and forward on purpose.
When removing the System.Drawing.Common the runtimeconfig.template.json file should also be removed.
A git sub-module of System.Windows.Forms was added in order to build the Linux example project.

---------

Co-authored-by: Randall Flagg <O@G>